Create Item class and create test for createJson()
