+++ src/test/java/com/google/edith/UserInsightsTest.java
@@ -9,7 +9,10 @@ import com.google.appengine.api.datastore.KeyFactory;
 import com.google.appengine.api.datastore.Query;
 import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
 import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
+import com.google.gson.Gson;
+import com.google.gson.JsonObject;
 import com.google.sps.servlets.UserInsights;
+import com.google.sps.servlets.Item;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
 import java.time.LocalDate;
@@ -158,4 +161,61 @@ public final class UserInsightsTest {
    
     Assert.assertEquals(map, userInsights.aggregateUserData());
   }
+
+//   @Test
+//   public void createJson() {    
+//     DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();   
+    
+//     List<Key> items = new ArrayList<>();
+//     List<Item> itemProperties = new ArrayList<>();
+
+//     items.add(KeyFactory.createKey("Item", "Item1"));
+//     items.add(KeyFactory.createKey("Item", "Item2"));
+//     items.add(KeyFactory.createKey("Item", "Item3"));
+//     items.add(KeyFactory.createKey("Item", "Item4"));
+
+//     Entity newEntity = new Entity(items.get(0));
+//     newEntity.setProperty("price", 5.00);
+//     newEntity.setProperty("quantity", 1);
+//     newEntity.setProperty("date", "2020-06-29");
+//     itemProperties.add(new Item(5.00, 1L, "2020-06-29"));
+//     datastore.put(newEntity); 
+
+//     Entity newEntity2 = new Entity(items.get(1));
+//     newEntity2.setProperty("price", 6.00);
+//     newEntity2.setProperty("quantity", 2);
+//     newEntity2.setProperty("date", "2020-06-30");
+//     itemProperties.add(new Item(6.00, 2L, "2020-06-30"));
+//     datastore.put(newEntity2); 
+    
+//     Entity newEntity3 = new Entity(items.get(2));
+//     newEntity3.setProperty("price", 7.00);
+//     newEntity3.setProperty("quantity", 3);
+//     newEntity3.setProperty("date", "2020-07-11");
+//     itemProperties.add(new Item(7.00, 3L, "2020-07-11"));
+//     datastore.put(newEntity3);  
+
+//     Entity newEntity4 = new Entity(items.get(3));
+//     newEntity4.setProperty("price", 8.00);
+//     newEntity4.setProperty("quantity", 4);
+//     newEntity4.setProperty("date", "2020-07-12");
+//     itemProperties.add(new Item(8.00, 4L, "2020-07-12"));
+//     datastore.put(newEntity4);  
+
+//     Map<String, Double> map = new HashMap<String, Double>();
+//     String mapJson = new Gson().toJson(map);
+//     map.put("2020-07-05", 17.00);
+//     map.put("2020-07-12", 53.00);
+//     userInsights.createUserStats();
+//     userInsights.updateUserStats(items);
+
+//     JsonObject testJson = new JsonObject();
+//     testJson.addProperty("weeklyAggregate", mapJson);
+//     String itemPropertiesJson = new Gson().toJson(items);
+//     testJson.addProperty("items", itemPropertiesJson);
+//     String expectedJson = new Gson().toJson(testJson);
+   
+//     Assert.assertEquals(expectedJson, userInsights.createJson());
+//   }
+
 }
